LayoutTests/crypto/subtle/ed25519-generate-export-key-raw-expected.txt

Test exporting an Ed25519 public key with raw format.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Generating a key pair...
Exporting the public key...
PASS publicKey.byteLength is 32
PASS successfullyParsed is true

TEST COMPLETE
